using strategy pattern to inject formFields
